fix: merge middleware into proxy for next.js 16
